refactor: Move item into single file
